feat(onClauseUpdated): implement onClauseUpdated and use hasErgoExpression to check if node should be replaced
