+++ package.json
@@ -63,10 +63,10 @@
     "webpack-dev-server": "^3.9.0"
   },
   "dependencies": {
-    "@accordproject/cicero-core": "^0.20.1",
-    "@accordproject/cicero-ui": "^0.2.4",
+    "@accordproject/cicero-core": "^0.20.3",
+    "@accordproject/cicero-ui": "^0.3.0",
     "@accordproject/ergo-compiler": "^0.20.1",
-    "@accordproject/markdown-slate": "^0.8.0",
+    "@accordproject/markdown-slate": "^0.8.2",
     "@babel/runtime": "^7.7.2",
     "mixin-deep": "^2.0.1",
     "monaco-editor": "^0.15.1",
